Issue 9038 Nested variable scope within lambdas
When lambdas reference variables declared in a
scope higher than the lambda itself was declared,
the .actualOuterLocalVariable binding we use
points to a variable that is not accessible
within the lambda.

Fortunately, the SyntheticArgumentBinding has the
same name as the correct LocalVariableBinding,
which we can retrieve from the MethodInfo object.

Change-Id: Ie202849960123037d5c3d3b52ff434aa26c99658