Add support for reflection in Gwt
This large commit has been developed and tested in the xapi project
at http://github.com/WeTheInternet/xapi

This adds support for three levels of reflection in Gwt: heavyweight,
lightweight and flyweight.

The heavyweight support will enhance a class
object with Methods, Fields, Constructors and Annotations, such that
runtime references to a class object will be able to perform reflection.
This is done using the GwtReflect.magicClass magic method, which will
enhance the entire object with all members (filtered using the annotation
ReflectionStrategy on the class in question).

Next is lightweight support, which is acheived via magic method on the
Class class itself.  Calling Class.getMethod() or .getField(), etc. on a class
literal, using literal strings and class array arguments will be translated
into a magic method call that instantiates only the single member
requested.  This is good for interop with other libraries, and for getting
access to a single member without bothering to enhance the entire class.

Finally, flyweight support is implemented via methods on the GwtReflect
class.  These methods will directly invoke a constructor or method, or
perform field get/set operations.   These magic methods require the class
containing the member, the string name of the member and (for executable
objects only,) the parameter class array to be a literal.  In addition, these
methods will also take an instance object and an (optional) array of
arguments.  The instance and arguments will be passed along to a method
which directly invokes method / constructor / field accessors. In a jvm,
these methods will perform normal reflection operations, so you can get
a minimum of overhead in Gwt without sacrificing JVM support.

Note that flyweight support may not be as optimized as possible yet.