Fix indeterministic jso member type resolution
When a class contains member types who extend each
other, and the supertype extends another type that
extends JavaScriptObject, GwtAstBuilder can accidentally
wind up setting the superclass of the subtype before the
subtype has its superclass set (and thus finds out that
it has a jso supertype), and does not realize that it (the
subtype) is actually a jso type.

This fix enforces an ordering which causes nested member
supertypes to be resolved before subtypes.