Allow default methods in gwt css interfaces.
Adds @IgnoreMethod as an annotation to skip an illegal method.

This lets you bake in other useful types that
don't actually know or care which resources you create;
handy for you to insert functionality,
and good for cross-platform code, which should not need
to reference gwt-specific resources interfaces
to know "I have some data I need as text/blob".