Preliminary Java 9 support
Note that this does not hook up the more relaxed try-with-resources
statements; we will add the new cases with tests in another commit.

Also note that our currently solution for classpath extraction
is quite messy; it worked in a toy project on a pre-release version
of java 9, before known changes to the module system (which we are
going to wait to settle before resuming work).