Add validation that Array.newInstance is not abused
Calling Array.newInstance(Type.class) without any int parameters
will call the multi-dimension varargs method, with a 0-length array.

This is illegal in java, so we force the generator to break as well,
since it would be worse to have something which works in gwt,
but is (correctly) broken in jvms.