Put back legacy ResourceLoaders methods
We put back the legacy "get URLClassLoader from thread" code,
which is not java 9 foolproofed, because we are responsible and
know that when manually calling these ClassLoader/Thread aware methods,
we need to be using a URLClassLoader for the compiler to be able to
extract the underlying classpath urls, for it's own lookup purposes.

Also, we aren't using java 9 yet, and when we do, we'll know that
we have to either ensure we are running in a thread using our
URLClassLoader, or choose another method to create and resolve
the ResourceLoader (by directly supplying a classpath of some kind).