Fix code and tests after rebase on 2.8.2
All dev tests except compiler test are passing,
with a laundry list in user; given that some of
the dev tests were failing on things we did not touch,
it is likely some / all of the user failures are
from upstream.

Also opens up more Recompiler methods to public use.